Add pipeline
Using CircleCI, as they provide (somewhat) easy to use and free pipeline
execution.

This pipeline runs basic validation and unit tests to verify commits.

It can also produce a release, a maintainer simply needs to tag the
desired commit with a semver tag (fx. 1.2.3) and push that tag. The zip
build of the plugin, which can be installed in QGIS, will be included in
the release.