Add pipeline (#3)
* Add pipeline

Using CircleCI, as they provide (somewhat) easy to use and free pipeline
execution.

This pipeline runs basic validation and unit tests to verify commits.

It can also produce a release, a maintainer simply needs to tag the
desired commit with a semver tag (fx. 1.2.3) and push that tag. The zip
build of the plugin, which can be installed in QGIS, will be included in
the release.

* Use proper context, as recommended

Recommended in this blog post:
https://circleci.com/blog/automating-your-releases-with-circleci-and-the-github-cli-orb/

The GITHUB_CREDS has been configured with a GITHUB_TOKEN environment
variable containing an access token issued by the dmi-cicd-bot that
never expires. This way, the pipeline is not dependent on any single DMI
employee.

Co-authored-by: Peter Garnæs <pga@dmi.dk>