Refactor, information tab now separate widget
The information UI config and code now lives in its own python module.
This is the first step in refactoring the main dialog code, and
separating out the code and UI config for each of the tabs. This will
improve separations of concern in the codebase, and improve readability.

Eventually run code (the code extracting data and importing it into QGIS
layers) for each of the tabs will live in the module of that tab. The
dialog class then simply becomes a controller of the tabs, relaying the
run operations.

This will impose some much needed structure of the code.