Merge pandas observation dataframes
Solves multiple issues. First of all, it deals with observations not of
the same temporal resolution, ie. the resulting table can now have empty
values in the parameter columns. When making a table of both a parameter
with 10 minute values and hourly values, this is a must.

Second of all, all station/parameter tables are rolled up into one big
table, so it can be exported to csv if desired by the user. Before only
the last station was exported.